Add middleware to set HTTP request content length
Adds middleware that will set the HTTP request's Content-Length header
if the length of the serialized stream can be determined. If the length
cannot be determined or the Header value is already set the middleware
will do nothing.

Fails the request if the underlying request stream reader is seekable
but fails to seek.