Fix protocol test generation for streaming types
Fixes the generation of protocol tests with streaming input and output
members. Refactors the comparison of structure values to use a utility
that flexibly compares the io.Reader values. Allowing the underlying
io.Reader implementations to be different types, but compare as equal if
the contents of the readers are the same.